+++ craft/commands/ssh_keys.py
@@ -1,7 +1,7 @@
 import click
 
 from craft.client import delete, get, post
-from craft.output import print_item, print_json, print_success, print_table
+from craft.output import print_item, print_success, print_table
 
 
 @click.command("list")
@@ -21,7 +21,7 @@ def sshkey_list():
             ])
         print_table(rows, ["ID", "Name", "Fingerprint"])
     else:
-        print_json(data)
+        print_item(data)
 
 
 @click.command("add")